refactor: share source scan utilities
